requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
pyyaml>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
